--- test/utils/shoots/common.go
@@ -1,20 +1,0 @@
-package shoots
-
-import (
-	"fmt"
-
-	gardencorev1beta1 "github.com/gardener/gardener/pkg/apis/core/v1beta1"
-)
-
-// ComputeTechnicalID computes the technical ID of a shoot
-func ComputeTechnicalID(projectName string, shoot *gardencorev1beta1.Shoot) string {
-	// Use the stored technical ID in the Shoot's status field if it's there.
-	// For backwards compatibility we keep the pattern as it was before we had to change it
-	// (double hyphens).
-	if len(shoot.Status.TechnicalID) > 0 {
-		return shoot.Status.TechnicalID
-	}
-
-	// New clusters shall be created with the new technical id (double hyphens).
-	return fmt.Sprintf("shoot--%s--%s", projectName, shoot.Name)
-}
